move treeview contextmenus to separate svelte components
