+++ client/src/ui/lib/treeView.svelte
@@ -1,336 +1,116 @@
 <script lang="ts">
-  import type { CreateGroup, EditGroup, DeleteGroup, ReorderGroup, CreateLayer, EditLayer, DeleteLayer, ReorderLayer } from '../../server/protocol'
+  import type { CreateGroup } from '../../server/protocol'
   import type { RenderMap } from '../../gl/renderMap'
-  import type { Color } from '../../twmap/types'
-  import type { Layer } from '../../twmap/layer'
-  import { TilesLayerFlags } from '../../twmap/types'
-  import { Image } from '../../twmap/image'
-  import { AnyTilesLayer, SwitchLayer, TuneLayer, FrontLayer, SpeedupLayer, TilesLayer, TeleLayer } from '../../twmap/tilesLayer'
+  import type { RenderLayer } from '../../gl/renderLayer'
   import { QuadsLayer } from '../../twmap/quadsLayer'
   import ContextMenu from './contextMenu.svelte'
-  import ImagePicker from './imagePicker.svelte'
-  import { decodePng, externalImageUrl, queryImage, isPhysicsLayer } from './util'
+  import GroupEditor from './editGroup.svelte'
+  import LayerEditor from './editLayer.svelte'
   import { showInfo, showError, clearDialog } from '../lib/dialog'
   import { server } from '../global'
 
   export let rmap: RenderMap
   export let visible = true
-  export let selected: [number, number] = [-1, -1]
+
+  export let g = -1
+  export let l = -1
+  $: rgroup = g !== -1 ? rmap.groups[g] : null
+  $: rlayer = rgroup && l !== -1 ? rgroup.layers[l] : null
 
   let folded = new Array(rmap.groups.length).fill(false)
 
-  function toStr(groupID: number, layerID: number) {
-    return `${groupID},${layerID}`
+  function toStr(g: number, l: number) {
+    return `${g},${l}`
   }
 
   function fromStr(str: string) {
     return str.split(',').map(x => parseInt(x)) as [number, number]
   }
 
-  function select(g: number, l: number) {
-    const [ oldG, oldL ] = selected
-    if (oldG !== -1 && oldL !== -1) {
-      const group = rmap.groups[oldG]
-      if (group) {
-        const layer = group.layers[oldL]
-        if (layer)
-          layer.active = false
-      }
-    }
-    if (g !== -1 && g !== -1) {
-      const group = rmap.groups[g]
-      if (group) {
-        const layer = group.layers[l]
-        if (layer)
-          layer.active = true
-      }
-    }
-    strSelected = toStr(g, l)
-    selected = [g, l]
-  }
-
-  let strSelected = toStr(...selected)
+  let strSelected = toStr(g, l)
   $: select(...fromStr(strSelected))
+
+  function select(new_g: number, new_l: number) {
+    if (rlayer)
+      rlayer.active = false
+
+    if (new_g !== -1 && new_l !== -1) {
+      const rlayer = rmap.groups[new_g].layers[new_l] // new_g and new_l are supposed to be valid
+      rlayer.active = true
+    }
+
+    strSelected = toStr(new_g, new_l)
+    g = new_g
+    l = new_l
+  }
   
   // ContextMenu
-  let cm = { g: null, l: null }
-  let cmX = 0
-  let cmY = 0
+  let cm_g = -1
+  let cm_l = -1
+  $: cm_rgroup = cm_g !== -1 ? rmap.groups[cm_g] : null
+  $: cm_rlayer = cm_rgroup && cm_l !== -1 ? rgroup.layers[cm_l] : null
+  let cm_x = 0
+  let cm_y = 0
 
-  function showCM(e: MouseEvent, g: number, l: number= null) {
-    cmX = e.clientX
-    cmY = e.clientY
-    cm = { g, l }
+  function showCM(e: MouseEvent, g: number, l: number = null) {
+    cm_x = e.clientX
+    cm_y = e.clientY
+    cm_g = g
+    cm_l = l
   }
 
   function hideCM() {
-    cm = { g: null, l: null }
+    cm_g = -1
+    cm_l = -1
   }
 
-  async function onEditLayer(change: EditLayer) {
-    try {
-      showInfo('Please wait…')
-      await server.query('editlayer', change)
-      rmap.editLayer(change)
-      rmap = rmap // hack to redraw treeview
-      clearDialog()
-    } catch (e) {
-      showError('Failed to edit layer: ' + e)
-    }
-  }
-  async function onEditGroup(change: EditGroup) {
-    try {
-      showInfo('Please wait…')
-      await server.query('editgroup', change)
-      rmap.editGroup(change)
-      rmap = rmap // hack to redraw treeview
-      clearDialog()
-    } catch (e) {
-      showError('Failed to edit group: ' + e)
-    }
-  }
   async function onCreateGroup() {
     try {
       const change: CreateGroup = { name: "" }
       showInfo('Please wait…')
       await server.query('creategroup', change)
       rmap.createGroup(change)
-      rmap = rmap // hack to redraw treeview
       clearDialog()
     } catch (e) {
       showError('Failed to create group: ' + e)
     }
   }
-  async function onCreateLayer(change: CreateLayer) {
-    try {
-      showInfo('Please wait…')
-      await server.query('createlayer', change)
-      rmap.createLayer(change)
-      rmap = rmap // hack to redraw treeview
-      clearDialog()
-    } catch (e) {
-      showError('Failed to create layer: ' + e)
-    }
-  }
-  async function onReorderGroup(change: ReorderGroup) {
-    try {
-      showInfo('Please wait…')
-      await server.query('reordergroup', change)
-      const [ g, l ] = selected
-      const group = rmap.groups[g]
-      rmap.reorderGroup(change)
-      select(rmap.groups.indexOf(group), l)
-      rmap = rmap // hack to redraw treeview
-      clearDialog()
-    } catch (e) {
-      showError('Failed to reorder group: ' + e)
-    }
-    hideCM()
-  }
-  async function onReorderLayer(change: ReorderLayer) {
-    try {
-      showInfo('Please wait…')
-      await server.query('reorderlayer', change)
-      const [ g, l ] = selected
-      const layer = rmap.groups[g].layers[l]
-      rmap.reorderLayer(change)
-      const newGroup = rmap.groups.find(g => g.layers.includes(layer))
-      select(rmap.groups.indexOf(newGroup), newGroup.layers.indexOf(layer))
-      rmap = rmap // hack to redraw treeview
-      clearDialog()
-    } catch (e) {
-      showError('Failed to reorder layer: ' + e)
-    }
-    hideCM()
-  }
-  async function onDeleteGroup(change: DeleteGroup) {
-    try {
-      showInfo('Please wait…')
-      await server.query('deletegroup', change)
-      const [ g, l ] = selected
-      const group = rmap.groups[g]
-      rmap.deleteGroup(change)
-      if (change.group === g) {
-        select(-1, -1)
-      }
-      else {
-        select(rmap.groups.indexOf(group), l)
-      }
-      rmap = rmap // hack to redraw treeview
-      clearDialog()
-    } catch (e) {
-      showError('Failed to delete group: ' + e)
-    }
-    hideCM()
-  }
-  async function onDeleteLayer(change: DeleteLayer) {
-    try {
-      showInfo('Please wait…')
-      await server.query('deletelayer', change)
-      const [ g, l ] = selected
-      const layer = rmap.groups[g].layers[l]
-      rmap.deleteLayer(change)
-      if (change.group === g && change.layer === l) {
-        select(-1, -1)
-      }
-      else {
-        const newGroup = rmap.groups.find(g => g.layers.includes(layer))
-        select(rmap.groups.indexOf(newGroup), newGroup.layers.indexOf(layer))
-      }
-      rmap = rmap // hack to redraw treeview
-      clearDialog()
-    } catch (e) {
-      showError('Failed to delete layer: ' + e)
-    }
-    hideCM()
-  }
 
-  function openFilePicker(g: number, l: number, layer: TilesLayer | QuadsLayer) {
-    const picker = new ImagePicker({
-      target: document.body,
-      props: {
-        images: rmap.map.images,
-        image: layer.image
-      },
-    })
-
-    picker.$on('pick', async (e: Event & { detail: Image | string | null }) => {
-      picker.$destroy()
-      const image = e.detail
-
-      if (image === null) { // no image used
-        onEditLayer({ group: g, layer: l, image: null })
-      }
-      else if (image instanceof Image) { // use embedded image
-        const index = rmap.map.images.indexOf(image)
-        onEditLayer({ group: g, layer: l, image: index })
-      }
-      else { // new external image
-        const index = rmap.map.images.length
-        const url = externalImageUrl(image)
-        const embed = await showInfo('Do you wish to embed this image?', 'yesno');
-        if (embed) {
-          try {
-            showInfo('Uploading image...', 'none')
-            const resp = await fetch(url)
-            const file = await resp.arrayBuffer()
-            await server.uploadFile(file)
-            await server.query('createimage', { name: image, index, external: false })
-            const img = await queryImage({ index })
-            rmap.addImage(img)
-            onEditLayer({ group: g, layer: l, image: index })
-            clearDialog()
-          }
-          catch (e) {
-            showError('Failed to upload image: ' + e)
-          }
-        }
-        else {
-          try {
-            showInfo('Creating image...', 'none')
-            const index = rmap.map.images.length
-            await server.query('createimage', { name: image, index, external: true })
-            const img = new Image()
-            img.loadExternal(url)
-            img.name = image
-            rmap.addImage(img)
-            onEditLayer({ group: g, layer: l, image: index })
-            clearDialog()
-          }
-          catch (e) {
-            showError('Failed to create external image: ' + e)
-          }
+  function layerIndex(rlayer: RenderLayer): [number, number] {
+    for (let g = 0; g < rmap.groups.length; g++) {
+      const rgroup = rmap.groups[g]
+      for (let l = 0; l < rgroup.layers.length; l++) {
+        if (rgroup.layers[l] === rlayer) {
+          return [ g, l ]
         }
       }
-    })
-
-    picker.$on('upload', async (e: Event & { detail: File }) => {
-      const image = e.detail
-      try {
-        showInfo('Uploading image...', 'none')
-        const name = image.name.replace(/\.[^\.]+$/, '')
-        const index = rmap.map.images.length
-        await server.uploadFile(await image.arrayBuffer())
-        await server.query('createimage', { name, index, external: false })
-        const data = await decodePng(image)
-        const img = new Image()
-        img.loadEmbedded(data)
-        img.name = name
-        rmap.addImage(img)
-        picker.$set({ images: rmap.map.images })
-        clearDialog()
-      }
-      catch (e) {
-        showError('Failed to upload image: ' + e)
-      }
-    })
-
-    picker.$on('delete', async (e: Event & { detail: Image }) => {
-      const image = e.detail
-
-      try {
-        const index = rmap.map.images.indexOf(image)
-        await server.query('deleteimage', { index })
-        rmap.removeImage(index)
-        picker.$set({ images: rmap.map.images })
-      }
-      catch (e) {
-        showError('Failed to delete image: ' + e)
-      }
-    })
-
-    picker.$on('cancel', () => {
-      picker.$destroy()
-    })
-  }
-
-  function colorToStr(c: Color) {
-    let hex = (i: number) => ('0' + i.toString(16)).slice(-2)
-    return `#${hex(c.r)}${hex(c.g)}${hex(c.b)}`
-  }
-
-  function strToColor(rgb: string, a: number) {
-    return {
-      r: parseInt(rgb.slice(1, 3), 16),
-      g: parseInt(rgb.slice(3, 5), 16),
-      b: parseInt(rgb.slice(5, 7), 16),
-      a,
     }
-  }
-  
-  function intVal(target: EventTarget) {
-    return parseInt((target as HTMLInputElement).value)
+
+    return [ -1, -1 ]
   }
 
-  function strVal(target: EventTarget) {
-    return (target as HTMLInputElement).value
-  }
-  
-  function layerName(layer: Layer) {
-    const quotedName = layer.name ? " '" + layer.name + "'" : ""
-    if (layer instanceof AnyTilesLayer) {
-      switch (layer.flags) {
-        case TilesLayerFlags.FRONT:
-          return "Front Layer"
-        case TilesLayerFlags.GAME:
-          return "Game Layer"
-        case TilesLayerFlags.SPEEDUP:
-          return "Speedup Layer"
-        case TilesLayerFlags.SWITCH:
-          return "Switch Layer"
-        case TilesLayerFlags.TELE:
-          return "Tele Layer"
-        case TilesLayerFlags.TILES:
-          return "Tile Layer" + quotedName
-        case TilesLayerFlags.TUNE:
-          return "Tune Layer"
+  function onChange() {
+    if (rlayer) {
+      const [ new_g, new_l ] = layerIndex(rlayer)
+      if (new_g !== g || new_l !== l) {
+        select(new_g, new_l)
+        rmap = rmap
       }
     }
-    else if (layer instanceof QuadsLayer) {
-      return "Quad Layer" + quotedName
+    if (cm_rlayer) {
+      const [ new_g, new_l ] = layerIndex(cm_rlayer)
+      if (new_g !== cm_g || new_l !== cm_l) {
+        cm_g = new_g
+        cm_l = new_l
+        rmap = rmap
+      }
     }
-    else {
-      return "Layer" + quotedName
+    else if (cm_rgroup) {
+      const new_g = rmap.groups.indexOf(cm_rgroup)
+      if (new_g !== g) {
+        cm_g = new_g
+        rmap = rmap
+      }
     }
   }
 
@@ -351,65 +131,9 @@
           <span class="options"
               on:click={(e) => showCM(e, g)}></span>
 
-          {#if cm.g === g && cm.l === null}
-            <ContextMenu x={cmX} y={cmY} on:close={hideCM}>
-              <span>Group #{g} {group.name}</span>
-              <label>Order <input type="number" min={0} max={rmap.groups.length - 1} value={g}
-                on:change={(e) => onReorderGroup({ group: g, newGroup: intVal(e.target) })}></label>
-              <label>Pos X <input type="number" value={group.offX}
-                on:change={(e) => onEditGroup({ group: g, offX: intVal(e.target) })}></label>
-              <label>Pos Y <input type="number" value={group.offY}
-                on:change={(e) => onEditGroup({ group: g, offY: intVal(e.target) })}></label>
-              <label>Para X <input type="number" value={group.paraX}
-                on:change={(e) => onEditGroup({ group: g, paraX: intVal(e.target) })}></label>
-              <label>Para Y <input type="number" value={group.paraY}
-                on:change={(e) => onEditGroup({ group: g, paraY: intVal(e.target) })}></label>
-              <label>Name <input type="text" value={group.name}
-                on:change={(e) => onEditGroup({ group: g, name: strVal(e.target) })}></label>
-              <button
-                on:click={() => onCreateLayer({ kind: 'tiles', group: g, name: "" })}>
-                Add tile layer
-              </button>
-              <button
-                on:click={() => onCreateLayer({ kind: 'quads', group: g, name: "" })}>
-                Add quad layer
-              </button>
-              {#if rgroup === rmap.physicsGroup}
-                {#if !rmap.map.physicsLayer(SwitchLayer)}
-                  <button
-                    on:click={() => onCreateLayer({ kind: 'switch', group: g, name: "" })}>
-                    Add switch layer
-                  </button>
-                {/if}
-                {#if !rmap.map.physicsLayer(FrontLayer)}
-                  <button
-                    on:click={() => onCreateLayer({ kind: 'front', group: g, name: "" })}>
-                    Add front layer
-                  </button>
-                {/if}
-                {#if !rmap.map.physicsLayer(TuneLayer)}
-                  <button
-                    on:click={() => onCreateLayer({ kind: 'tune', group: g, name: "" })}>
-                    Add tune layer
-                  </button>
-                {/if}
-                {#if !rmap.map.physicsLayer(SpeedupLayer)}
-                  <button
-                    on:click={() => onCreateLayer({ kind: 'speedup', group: g, name: "" })}>
-                    Add speedup layer
-                  </button>
-                {/if}
-                {#if !rmap.map.physicsLayer(TeleLayer)}
-                  <button
-                    on:click={() => onCreateLayer({ kind: 'tele', group: g, name: "" })}>
-                    Add tele layer
-                  </button>
-                {/if}
-              {/if}
-              <button
-                on:click={() => onDeleteGroup({ group: g })}>
-                Delete group
-              </button>
+          {#if cm_g === g && cm_l === -1}
+            <ContextMenu x={cm_x} y={cm_y} on:close={hideCM}>
+              <GroupEditor {rmap} {g} on:change={onChange} />
             </ContextMenu>
           {/if}
 
@@ -427,45 +151,9 @@
             <span class="options"
               on:click={(e) => showCM(e, g, l)}></span>
 
-            {#if cm.g === g && cm.l === l}
-              <ContextMenu x={cmX} y={cmY} on:close={hideCM}>
-                <span>{layerName(layer)}</span>
-                {#if !isPhysicsLayer(layer)}
-                  <label>Group <input type="number" min={0} max={rmap.groups.length - 1} value={g}
-                    on:change={(e) => onReorderLayer({ group: g, layer: l, newGroup: intVal(e.target), newLayer: 0 })}></label>
-                {/if}
-                <label>Order <input type="number" min={0} max={group.layers.length - 1} value={l}
-                  on:change={(e) => onReorderLayer({ group: g, layer: l, newGroup: g, newLayer: intVal(e.target) })}></label>
-                {#if layer instanceof TilesLayer}
-                  <label>Width <input type="number" min={2} max={10000} value={layer.width}
-                    on:change={(e) => onEditLayer({ group: g, layer: l, width: intVal(e.target) })}></label>
-                  <label>Height <input type="number" min={2} max={10000} value={layer.height}
-                    on:change={(e) => onEditLayer({ group: g, layer: l, height: intVal(e.target) })}></label>
-                  {#if layer instanceof TilesLayer}
-                    {@const img = layer.image ? layer.image.name : "<none>" }
-                    <label>Image <input type="button" value={img}
-                      on:click={() => openFilePicker(g, l, layer)}></label>
-                    {@const col = layer.color}
-                    <label>Color <input type="color" value={colorToStr(layer.color)}
-                      on:change={(e) => onEditLayer({ group: g, layer: l, color: strToColor(strVal(e.target), col.a) })}></label>
-                    <label>Opacity <input type="range" min={0} max={255} value={col.a}
-                      on:change={(e) => onEditLayer({ group: g, layer: l, color: { ...col, a: intVal(e.target) } })}></label>
-                    {#if layer.flags === TilesLayerFlags.TILES}
-                      <label>Name <input type="text" value={layer.name}
-                        on:change={(e) => onEditLayer({ group: g, layer: l, name: strVal(e.target) })}></label>
-                    {/if}
-                  {/if}
-                {:else if layer instanceof QuadsLayer}
-                  {@const img = layer.image ? layer.image.name : "<none>" }
-                  <label>Image <input type="button" value={img}
-                    on:click={() => openFilePicker(g, l, layer)}></label>
-                  <label>Name <input type="text" value={layer.name}
-                    on:change={(e) => onEditLayer({ group: g, layer: l, name: strVal(e.target) })}></label>
-                {/if}
-                <button
-                  on:click={() => onDeleteLayer({ group: g, layer: l })}>
-                  Delete layer
-                </button>
+            {#if cm_g === g && cm_l === l}
+              <ContextMenu x={cm_x} y={cm_y} on:close={hideCM}>
+                <LayerEditor {rmap} {g} {l} on:change={onChange} />
               </ContextMenu>
             {/if}
